Increase optimization in Linaro's /Dalvik
Android 4.1-linaro changed this to -fno-inline-fucntions to solve a
boot loop problem compiling under -O3 with GCC 4.6+.  -fno-inline-functions
is an override for optimization CFLAGS used to help determine the cause of
bugs speculated to be on the GCC side. So for now.. lets restore -O2 since we don't have hair
that long.

PS2: WTFBBQZOMG

Change-Id: Ic46b072d30c0b53ac441e44e2a05110b611cc51e